Patch tests to use a common setup and teardown method
